moving the package json util to utils
